Restructured project to met-cron
