+++ met-cron/.idea/.gitignore
@@ -1,0 +1,4 @@
+# Default ignored files
+/shelf/
+/workspace.xml
+/.name
