+++ met-cron/pre-hook-update-db.sh
