+++ met-cron/setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = met-etl
+name = met-cron
 url = https://github.com/bcgov/met-public/
 author = MET Team
 author_email =
@@ -46,7 +46,7 @@ max_line_length = 120
 ignore = E501
 docstring-min-length=10
 notes=FIXME,XXX # TODO is ignored
-match_dir = src/met_etl
+match_dir = src/met_cron
 ignored-modules=flask_sqlalchemy
     sqlalchemy
 per-file-ignores =
