+++ met-cron/setup.py
@@ -47,7 +47,7 @@ def read(filepath):
 REQUIREMENTS = read_requirements('requirements/prod.txt')
 
 setup(
-    name="met_etl",
+    name="met_cron",
     packages=find_packages('src'),
     package_dir={'': 'src'},
     py_modules=[splitext(basename(path))[0] for path in glob('src/*.py')],
