+++ openshift/redash_bc.yml
@@ -1,0 +1,60 @@
+apiVersion: template.openshift.io/v1
+kind: Template
+apiVersion: template.openshift.io/v1
+metadata:
+  name: forms-flow-analytics
+  annotations:
+    description: ""
+    tags: forms-flow-analytics
+parameters:
+  - name: SOURCE_REPOSITORY_URL
+    displayName: Git Repository URL
+    description: The URL of the repository with your application source code.
+    value: https://github.com/VineetBala-AOT/redash
+  - name: SOURCE_REPOSITORY_REF
+    displayName: Git Reference
+    description: Set this to a branch name, tag or other ref of your repository if you are not using the default branch.
+    value: master
+  - name: SOURCE_CONTEXT_DIR
+    displayName: Git context directory
+    description: Set this to a branch name, tag or other ref of your repository if you are not using the default branch.
+objects:
+-
+  apiVersion: build.openshift.io/v1
+  kind: BuildConfig
+  metadata:
+    name: forms-flow-analytics
+    labels:
+      app: forms-flow-analytics
+  spec:
+    resources:
+      limits:
+        cpu: '2'
+        memory: 8Gi
+      requests:
+        cpu: '1'
+        memory: 4Gi
+    source:
+      git:
+        ref: ${SOURCE_REPOSITORY_REF}
+        uri: ${SOURCE_REPOSITORY_URL}
+    output:
+      to:
+        kind: ImageStreamTag
+        name: 'forms-flow-analytics:latest'
+    strategy:
+      type: Docker
+      dockerStrategy:
+        pullSecret:
+          name: docker-creds
+      dockerfilePath: Dockerfile
+    triggers:
+      - type: ImageChange
+        imageChange: {}
+      - type: ConfigChange
+
+-
+  apiVersion: image.openshift.io/v1
+  kind: ImageStream
+  metadata:
+    name: forms-flow-analytics
